requirements.txt
asgiref==3.2.7
Django==3.0
django-bootstrap-modal-forms==1.5.0
django-crispy-forms==1.9.0
django-filters==0.2.1
django-filter==2.2.0
django-widget-tweaks==1.4.8
gunicorn==20.0.4
Pillow==7.1.2
pytz==2020.1
sqlparse==0.3.1
whitenoise==5.0.1
